Move remaining activities -> `kopf.reactor.activities`
