+++ kopf/engines/probing.py
@@ -6,8 +6,8 @@ from typing import Optional, Tuple, MutableMapping
 
 import aiohttp.web
 
+from kopf.reactor import activities
 from kopf.reactor import causation
-from kopf.reactor import handling
 from kopf.reactor import lifecycles
 from kopf.reactor import registries
 
@@ -50,7 +50,7 @@ async def health_reporter(
                 now = datetime.datetime.utcnow()
                 if probing_timestamp is None or now - probing_timestamp >= probing_max_age:
 
-                    activity_results = await handling.run_activity(
+                    activity_results = await activities.run_activity(
                         lifecycle=lifecycles.all_at_once,
                         registry=registry,
                         activity=causation.Activity.PROBE,
