+++ tests/handling/test_activity_triggering.py
@@ -3,8 +3,9 @@ from typing import Mapping
 import freezegun
 import pytest
 
+from kopf.reactor.activities import ActivityError, run_activity
 from kopf.reactor.causation import Activity
-from kopf.reactor.handling import ActivityError, PermanentError, TemporaryError, run_activity
+from kopf.reactor.handling import PermanentError, TemporaryError
 from kopf.reactor.lifecycles import all_at_once
 from kopf.reactor.registries import HandlerId, OperatorRegistry
 from kopf.reactor.states import HandlerOutcome
